third_party/blink/web_tests/fast/frames/avoid-forced-layout-on-empty-subframe-expected.txt

Test that inserting an empty iframe doesn't force a layout

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS layoutCountAfter is layoutCountBefore
PASS successfullyParsed is true

TEST COMPLETE
